+++ README.md
@@ -2,7 +2,7 @@
 Final Project for 6.C85, Interactive Data Visualization &amp; Society. This project focuses on the remittances associated with migration from the Northern Triangle and if that money sent back offsets the cost of migration.
 
 # Data
-World Bank data is available in the `remit-return/remitreturn/wb-data/` directory. Class data must remain offline (private). Download the folder `dataset-1_central-american-survey` locally to your computer and place it in the `remit-return/remitreturn/class-data/` folder. Then you can run the data prep files.
+World Bank data is available in the `remit-return/remitreturn/wb-data/` directory. Class data must remain offline (private). Download the folder `dataset-1_central-american-survey` locally to your computer and place it in the `remit-return/remitreturn/src/class-data/` folder. Then you can run the data prep files.
 
 # Local
 ## Running the code
