+++ remitreturn/class-data/migration_motivations.csv
@@ -1,4 +1,4 @@
-﻿group,var1,var2,var3,total
+﻿motivName,var1,var2,var3,total
 Better Salary,412,627,746,1785
 Unemployment,151,271,238,660
 Covering basic needs,158,150,218,526
