+++ remitreturn/class-data/migration_motivations.js
@@ -1,104 +1,104 @@
 export let migration_motivations = [
     {
-        group: "Better Salary",
+        motivName: "Better Salary",
         GTM: 412,
         HND: 627,
         SLV: 746,
         total: 1785
     },
     {
-        group: "Unemployment",
+        motivName: "Unemployment",
         GTM: 151,
         HND: 271,
         SLV: 238,
         total: 660
     },
     {
-        group: "Covering basic needs",
+        motivName: "Covering basic needs",
         GTM: 158,
         HND: 150,
         SLV: 218,
         total: 526
     },
     {
-        group: "Remittances",
+        motivName: "Remittances",
         GTM: 87,
         HND: 118,
         SLV: 180,
         total: 385
     },
     {
-        group: "Lack of money for food",
+        motivName: "Lack of money for food",
         GTM: 92,
         HND: 178,
         SLV: 100,
         total: 370
     },
     {
-        group: "Tourism",
+        motivName: "Tourism",
         GTM: 29,
         HND: 42,
         SLV: 86,
         total: 157
     },
     {
-        group: "Family reunification",
+        motivName: "Family reunification",
         GTM: 23,
         HND: 35,
         SLV: 75,
         total: 133
     },
     {
-        group: "Natural hazard",
+        motivName: "Natural hazard",
         GTM: 14,
         HND: 68,
         SLV: 42,
         total: 124
     },
     {
-        group: "Unsafety",
+        motivName: "Unsafety",
         GTM: 5,
         HND: 33,
         SLV: 61,
         total: 99
     },
     {
-        group: "Other",
+        motivName: "Other",
         GTM: 16,
         HND: 28,
         SLV: 23,
         total: 67
     },
     {
-        group: "Study",
+        motivName: "Study",
         GTM: 15,
         HND: 29,
         SLV: 18,
         total: 62
     },
     {
-        group: "Cultural reasons",
+        motivName: "Cultural reasons",
         GTM: 0,
         HND: 12,
         SLV: 18,
         total: 30
     },
     {
-        group: "Health",
+        motivName: "Health",
         GTM: 4,
         HND: 11,
         SLV: 12,
         total: 27
     },
     {
-        group: "Loss of land",
+        motivName: "Loss of land",
         GTM: 5,
         HND: 4,
         SLV: 0,
         total: 9
     },
     {
-        group: "Domestic violence",
+        motivName: "Domestic violence",
         GTM: 2,
         HND: 1,
         SLV: 2,
