+++ remitreturn/src/components/bar.svelte
@@ -1,5 +1,112 @@
 <script>
-  import { migration_motivations } from "../../../class-data/migration_motivations.js";
+  import { migration_motivations } from "../../class-data/migration_motivations.js";
+  /*const migration_motivations = [
+    {
+      motivName: "Better Salary",
+      GTM: 412,
+      HND: 627,
+      SLV: 746,
+      total: 1785,
+    },
+    {
+      motivName: "Unemployment",
+      GTM: 151,
+      HND: 271,
+      SLV: 238,
+      total: 660,
+    },
+    {
+      motivName: "Covering basic needs",
+      GTM: 158,
+      HND: 150,
+      SLV: 218,
+      total: 526,
+    },
+    {
+      motivName: "Remittances",
+      GTM: 87,
+      HND: 118,
+      SLV: 180,
+      total: 385,
+    },
+    {
+      motivName: "Lack of money for food",
+      GTM: 92,
+      HND: 178,
+      SLV: 100,
+      total: 370,
+    },
+    {
+      motivName: "Tourism",
+      GTM: 29,
+      HND: 42,
+      SLV: 86,
+      total: 157,
+    },
+    {
+      motivName: "Family reunification",
+      GTM: 23,
+      HND: 35,
+      SLV: 75,
+      total: 133,
+    },
+    {
+      motivName: "Natural hazard",
+      GTM: 14,
+      HND: 68,
+      SLV: 42,
+      total: 124,
+    },
+    {
+      motivName: "Unsafety",
+      GTM: 5,
+      HND: 33,
+      SLV: 61,
+      total: 99,
+    },
+    {
+      motivName: "Other",
+      GTM: 16,
+      HND: 28,
+      SLV: 23,
+      total: 67,
+    },
+    {
+      motivName: "Study",
+      GTM: 15,
+      HND: 29,
+      SLV: 18,
+      total: 62,
+    },
+    {
+      motivName: "Cultural reasons",
+      GTM: 0,
+      HND: 12,
+      SLV: 18,
+      total: 30,
+    },
+    {
+      motivName: "Health",
+      GTM: 4,
+      HND: 11,
+      SLV: 12,
+      total: 27,
+    },
+    {
+      motivName: "Loss of land",
+      GTM: 5,
+      HND: 4,
+      SLV: 0,
+      total: 9,
+    },
+    {
+      motivName: "Domestic violence",
+      GTM: 2,
+      HND: 1,
+      SLV: 2,
+      total: 5,
+    },
+  ];*/
 
   export let countrySelection = "all";
   export let themeColors;
@@ -48,7 +155,7 @@
             text-anchor="end"
             x={labelPositionXLeft}
             y={labelPositionY}
-            >{motivation.group}
+            >{motivation.motivName}
           </text>
           <text
             class="motivation-bar-chart-label"
@@ -71,7 +178,7 @@
             text-anchor="end"
             class="motivation-bar-chart-label"
             x={labelPositionXLeft}
-            y={labelPositionY}>{motivation.group}</text
+            y={labelPositionY}>{motivation.motivName}</text
           >
           <text
             class="motivation-bar-chart-label"
@@ -94,7 +201,7 @@
             class="motivation-bar-chart-label"
             text-anchor="end"
             x={labelPositionXLeft}
-            y={labelPositionY}>{motivation.group}</text
+            y={labelPositionY}>{motivation.motivName}</text
           >
           <text
             class="motivation-bar-chart-label"
@@ -117,7 +224,7 @@
             class="motivation-bar-chart-label"
             text-anchor="end"
             x={labelPositionXLeft}
-            y={labelPositionY}>{motivation.group}</text
+            y={labelPositionY}>{motivation.motivName}</text
           >
           <text
             class="motivation-bar-chart-label"
